change: use erofs image as nix-store
This image nix-store has only the packages/derivations needed for the
VM. The VM doesn't need access to all packages in the hosts nix-store.

This comes at cost: the nix-store image is per-VM and can be quite
large (a few GB). It's stored in the hosts nix-store.

An alternative would have been virtualisation.useNixStoreImage, but
there the store image is build at runtime, rather that at build-time,
drastically increasing VM start-up time.

based on and inspried by:
- https://git.scottworley.com/nixos-qemu-vm-isolation
- https://github.com/chkno/nixos-qemu-vm-isolation